+++ packages/browser/src/canvas/convert.ts
@@ -1,53 +1,52 @@
-import { getElementSize, getSize as getSizeImage } from "../image/utils";
-import { Size } from "@a2-ts-utils/common/type";
+import { ImageMimeType, Size } from "@a2-ts-utils/common/type";
+import { getElementSize, getSize as getSizeImage } from "../image/image";
 
 // elementからCanvasを取得
 // newSizeを指定した場合は、Canvasのサイズを変更して描画する
-export function elementToCanvas(
+export function elementToCanvas2d(
   element: HTMLVideoElement | HTMLImageElement | HTMLCanvasElement,
   newSize?: Size,
 ): HTMLCanvasElement {
   const canvas = document.createElement("canvas");
 
-  if (newSize) {
-    canvas.width = newSize.width;
-    canvas.height = newSize.height;
-  } else {
-    const size = getElementSize(element);
-    canvas.width = size.width;
-    canvas.height = size.height;
-  }
+  const _size = newSize ?? getElementSize(element);
+  canvas.width = Math.floor(_size.width);
+  canvas.height = Math.floor(_size.height);
 
   const ctx = canvas.getContext("2d");
   if (!ctx) {
     throw new Error("Failed to get canvas context");
   }
+
+  ctx.imageSmoothingEnabled = true;
+  ctx.imageSmoothingQuality = "high";
   ctx.drawImage(element, 0, 0, canvas.width, canvas.height);
 
   return canvas;
 }
 
-export const ToDataUrlType = {
-  JPEG: "image/jpeg",
-  PNG: "image/png",
-  WEBP: "image/webp",
-  AVIF: "image/avif",
-} as const;
-export type ToDataUrlType = (typeof ToDataUrlType)[keyof typeof ToDataUrlType];
-
-export async function getBytes(
+// Blob: Binary Large Object
+// canvas > Blob > buffer > Uint8Array
+export function getBlobAsync(
   canvas: HTMLCanvasElement,
-  newType: ToDataUrlType,
-): Promise<Uint8Array> {
+  type: ImageMimeType = ImageMimeType.JPEG,
+): Promise<Blob> {
   return new Promise((resolve, reject) => {
     canvas.toBlob((blob) => {
-      if (!blob) {
-        reject(new Error("Failed to create blob"));
-        return;
-      }
-      blob.arrayBuffer().then((buffer) => {
-        resolve(new Uint8Array(buffer));
-      });
-    }, newType);
+      blob ? resolve(blob) : reject(new Error("Failed to create blob"));
+    }, type);
   });
+}
+
+export async function blobToBytesAsync(blob: Blob): Promise<Uint8Array> {
+  const buffer = await blob.arrayBuffer();
+  return new Uint8Array(buffer);
+}
+
+export async function getBytesAsync(
+  canvas: HTMLCanvasElement,
+  type: ImageMimeType = ImageMimeType.JPEG,
+): Promise<Uint8Array> {
+  const blob = await getBlobAsync(canvas, type);
+  return blobToBytesAsync(blob);
 }