+++ packages/browser/src/image/index.ts
@@ -1,19 +1,2 @@
-export type ImageInfo = {
-  src: string;
-  // 元サイズ
-  naturalWidth: number;
-  naturalHeight: number;
-  // 表示サイズ
-  width: number;
-  height: number;
-};
-
-export function getImageInfo(element: HTMLImageElement): ImageInfo {
-  return {
-    src: element.src,
-    naturalWidth: element.naturalWidth,
-    naturalHeight: element.naturalHeight,
-    width: element.width,
-    height: element.height,
-  };
-}
+export * from "./image";
+export * from "./convert";