+++ packages/browser/history.md
@@ -2,4 +2,6 @@
 
 ## 26-04-29(wed)
 
+- 画像の変換処理を分割及びカスタマイズ
+  - element > canvas > blob > buffer > bytes
 - VideoやImageのリサイズ処理を追加