+++ packages/common/history.md
@@ -2,5 +2,6 @@
 
 ## 26-04-29(wed)
 
+- ImageMimeType を追加
 - 画像用の関数やtypeを追加
   - getFitSize, Size など